move row and filter-row to components level
